Move logit methods into separate shared module.
